Senesor error format fix (#1111)
* Sensors: Fix formatting error, extra placeholders

Few sensors had similar format for errors that did not match
argument list.
This change removes extra placeholders.

* Sensors: Fix formatting for LOG_ERROR calls

In few cases format list did not match arguments.
This could cause warnings if compiler would check for
printf like inconsistency.
Warning will popup when log_printf() will be change to have:
__attribute__ ((format (printf, 4, 5)));

+++ hw/drivers/sensors/bno055/src/bno055.c
@@ -889,7 +889,7 @@ bno055_find_reg(sensor_type_t type, uint8_t *reg)
             *reg = BNO055_GRAVITY_DATA_X_LSB_ADDR;
             break;
         default:
-            BNO055_ERR("Not supported sensor type: %d\n", type);
+            BNO055_ERR("Not supported sensor type: %d\n", (int)type);
             rc = SYS_EINVAL;
             break;
     }
